```text Refactor ClickValue component and remove unused props
The ClickValue component was refactored to remove the unused props. The props object was removed from the component definition.

Add form_controller_patch.js for handling rewards

A new file, form_controller_patch.js, was added to handle rewards in the form controller. This file patches the FormController prototype to add a 1% chance of giving a reward to the user. If a reward is won, a notification is displayed with the reward description and a "Collect" button. Clicking the button applies the reward and opens a new client action for the Clicker Game.

Update res_users_view_form.xml

The res_users_view_form.xml file was updated to remove the one2many_list widget from the property_ids field in the Properties page of the notebook.